# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AcmeX API
  description: Enterprise ACME v2 management API for automated TLS certificate lifecycle.
  version: 0.9.0
servers:
  - url: http://localhost:8080/api/v1
    description: Certificate lifecycle API v1
paths:
  /challenge-cleanup:
    get:
      summary: List challenge leases awaiting cleanup (including cleanup_failed)
      description: >
        Operator cleanup queue (T05): active, cleanup_pending and
        cleanup_failed leases as redacted views. cleanup_failed leases are
        the target of the manual retry entry.
      parameters:
        - name: pending
          in: query
          description: >
            Only the pending cleanup queue is exposed in v1; false is
            rejected with a problem response.
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Cleanup-pending lease list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChallengeLeaseView'
        '400':
          $ref: '#/components/responses/Problem'
...
  /certificate-intents:
    get:
      summary: List certificate intents
      responses:
        '200':
          description: Intent list
    post:
      summary: Create a certificate intent
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIntentRequest'
      responses:
        '201':
          description: Intent created or idempotently replayed
        '400':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
...
  responses:
...
    Problem:
      description: RFC 7807 problem response with AcmeX stable error code
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    CreateIntentRequest:
      type: object
      required:
        - identifiers
      properties:
        identifiers:
          type: array
          minItems: 1
          items:
            type: string
        ca_policy:
          type: object
        validation_policy:
          type: object
        key_policy:
          type: object
        renewal_policy:
          type: object
        delivery_targets:
          type: array
          items:
            type: object
        external_csr:
          type: string
          description: >
            PEM-encoded PKCS#10 CSR (`CERTIFICATE REQUEST`) held by the
            caller. Required when `key_policy.mode` is `external_csr` and
            forbidden when it is `managed` (violations are rejected with
            400). AcmeX never derives or stores the private key; the CSR is
            validated (signature + exact SAN match) and forwarded to the CA
            at finalize.
...
    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - error_code
        - retryable
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        error_code:
          type: string
        retryable:
          type: boolean
        operation_id:
          type: string
...
    ChallengeLeaseView:
      type: object
      description: >
        Redacted challenge lease projection for the cleanup queue (T05):
        the locator is collapsed to a credential-free summary (DNS zone +
        record name, HTTP/TLS agent + route ids).
      required:
        - id
        - operation_id
        - challenge_type
        - locator_summary
        - cleanup_attempts
        - state
      properties:
        id:
          type: string
        operation_id:
          type: string
        challenge_type:
          type: string
          enum: [dns-01, http-01, tls-alpn-01]
        locator_summary:
          type: string
        cleanup_attempts:
          type: integer
          format: uint32
          minimum: 0
        last_cleanup_error:
          type: string
        state:
          type: string
          enum: [active, cleanup_pending, cleaned, cleanup_failed]
      not:
        required:
          - token
          - key_authorization
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - ApiKeyAuth: []